requirements.txt
pika==1.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
et-xmlfile==1.1.0
idna==3.3
numpy==1.21.4
openpyxl==3.0.9
pandas==1.3.4
psycopg2-binary==2.9.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
tqdm==4.62.3
urllib3==1.26.7
